kdrive: Fix tslib check fallback to set TSLIB_LIBS. #30599
If pkg-config is unable to find tslib but the fallback check does find
it, the compilation continues with tslib support enabled though
TSLIB_LIBS are unset. Thus, the compilation fails with a linking error
on tslib functions.

This patch sets TSLIB_LIBS to '-lts' whenever the tslib fallback check
succeeds.

Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=30599
Signed-off-by: Michał Górny <mgorny@gentoo.org>
Reviewed-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Keith Packard <keithp@keithp.com>